fix micropython includes

+++ scripts/build-micropython.sh
@@ -63,10 +63,13 @@ fi
 # in order to build; these need to end up in top level include directory
 # so that they are found by compiler/assembler.
 #
-LITEX_INCLUDE_BASE="$PWD/third_party/litex/litex/soc/software/include/base"
+LITEX_INCLUDE_BASE="$PWD/third_party/litex/litex/soc/cores/cpu/$CPU"
 
 for FILE in system.h csr-defs.h spr-defs.h; do
-	cp -p "$LITEX_INCLUDE_BASE/$FILE" "$TARGET_BUILD_DIR/software/include"
+        # spr-defs.h is available only for selected CPUs
+        if [ -e "$LITEX_INCLUDE_BASE/$FILE" ]; then
+        	cp -p "$LITEX_INCLUDE_BASE/$FILE" "$TARGET_BUILD_DIR/software/include"
+    	fi
 done
 
 # Setup the micropython build directory
